PrevRunInfo moved to a new module prev_run.py, and cleanup
